Refactor utility functions
